applying the night mode to the map section component, changing broken url videos
