+++ client/src/components/DotMap.css
@@ -18,5 +18,5 @@
 }
 
 .night #dotmap path {
-	fill: gold;
+	fill: rgb(185, 158, 1);
 }
